+++ CHANGELOG.md
@@ -12,6 +12,7 @@
   - AirPortBrcmFixup 2.1.6 - release
   - DebugEnhancer 1.0.7 - release
   - FeatureUnlock 1.1.0 - rolling (2bb3c50)
+  - RestrictEvents 1.0.9 - rolling (f3db795)
 
 ## 0.5.1
 - Add support for `APPLE SSD TS0128F/256F` SSDs in macOS Ventura